Implementato spostamento paginazione gallery tramite Slider. Quando viene spostato lo Slider, viene aperto un Popup con il numero del fotogramma della prima foto di quella pagina.
Eliminate alcune variabili unsigned perch? non sono utili e non sono tipiche
del linguaggio c# ma servono solo per compatibilit? con il linguaggio C++

Modificati i numeri di versione.

+++ Digiphoto.Lumen.UI/Dialogs/StampaProviniDialog.xaml.cs
@@ -170,8 +170,8 @@ namespace Digiphoto.Lumen.UI.Dialogs
 		private void RapidoButton_Click(object sender, RoutedEventArgs e) {
 			String tag = (String) (e.Source as Button).Tag;
 			var vetRC = tag.Split('x');
-			model.paramStampaProvini.numeroRighe = int.Parse(vetRC[0]);
-			model.paramStampaProvini.numeroColonne = int.Parse(vetRC[1]);
+			model.paramStampaProvini.numeroRighe = short.Parse(vetRC[0]);
+			model.paramStampaProvini.numeroColonne = short.Parse(vetRC[1]);
 		}
 	}
 }
